@@ -1,22 +1,22 @@
-import React from 'react';
+import React from "react";
 
-import Banner, { BannerPassthroughProps } from './Banner';
-import { LinkedInButton, TwitterButton } from './SocialButton';
+import Banner, { BannerPassthroughProps } from "./Banner";
+import { LinkedInButton, TwitterButton } from "./SocialButton";
 
 export interface HeroProps extends BannerPassthroughProps {
   /** Name of our protagonist */
-  name: string,
+  name: string;
   /** URL to a headshot of the protagonist */
-  headshot: string,
+  headshot: string;
 }
 
 interface HeroState {
-  height: number,
+  height: number;
 }
 
 /**
  * A hero banner
- * 
+ *
  * Contains an image and two speech bubbles
  */
 export default class Hero extends React.Component<HeroProps, HeroState> {
@@ -26,7 +26,7 @@ export default class Hero extends React.Component<HeroProps, HeroState> {
   static MIN_ANIM_WINDOW_WIDTH = 900;
   static MIN_ANIM_WINDOW_HEIGHT = 700;
 
-  readonly state:HeroState = {
+  readonly state: HeroState = {
     height: Hero.computeHeight(),
   };
 
@@ -34,10 +34,10 @@ export default class Hero extends React.Component<HeroProps, HeroState> {
     super(props);
     this.setHeight = this.setHeight.bind(this);
   }
-  
+
   componentDidMount() {
-    document.addEventListener('scroll', this.setHeight);
-    window.addEventListener('resize', this.setHeight);
+    document.addEventListener("scroll", this.setHeight);
+    window.addEventListener("resize", this.setHeight);
 
     // As we use react-snap and ReactDOM.hydrate in prod, if the user refreshes
     // the page while scrolled down the .hero-content max-height will mismatch.
@@ -49,17 +49,31 @@ export default class Hero extends React.Component<HeroProps, HeroState> {
   }
 
   componentWillUnmount() {
-    document.removeEventListener('scroll', this.setHeight);
-    window.removeEventListener('resize', this.setHeight);
+    document.removeEventListener("scroll", this.setHeight);
+    window.removeEventListener("resize", this.setHeight);
   }
 
   static computeHeight(): number {
     // Match the min-width and min-height media rules for applying a header animation
-    if (window.innerWidth < Hero.MIN_ANIM_WINDOW_WIDTH || window.innerHeight < Hero.MIN_ANIM_WINDOW_HEIGHT) {
+    if (
+      window.innerWidth < Hero.MIN_ANIM_WINDOW_WIDTH ||
+      window.innerHeight < Hero.MIN_ANIM_WINDOW_HEIGHT
+    ) {
+      console.log("1", Hero.MAX_HEIGHT);
       return Hero.MAX_HEIGHT;
     } else {
       const top = window.scrollY;
-      return top < Hero.SCROLL_CUTOFF ? Hero.MAX_HEIGHT - top : Hero.MAX_HEIGHT - Hero.SCROLL_CUTOFF;
+      console.log(
+        "2",
+        top,
+        top < Hero.SCROLL_CUTOFF,
+        top < Hero.SCROLL_CUTOFF
+          ? Hero.MAX_HEIGHT - top
+          : Hero.MAX_HEIGHT - Hero.SCROLL_CUTOFF
+      );
+      return top < Hero.SCROLL_CUTOFF
+        ? Hero.MAX_HEIGHT - top
+        : Hero.MAX_HEIGHT - Hero.SCROLL_CUTOFF;
     }
   }
 
@@ -73,19 +87,23 @@ export default class Hero extends React.Component<HeroProps, HeroState> {
     const headshotSize = height * 0.8;
 
     return (
-      <Banner className='hero' {...rest}>
-        <div className='hero-content' style={{ maxHeight: height }}>
-          <img src={headshot} alt={name} style={{
-            width: headshotSize,
-            height: headshotSize,
-          }}/>
-          <div className='conversation'>
-            <div className='greeting speech-bubble'>
+      <Banner className="hero" {...rest}>
+        <div className="hero-content" style={{ maxHeight: Math.floor(height) }}>
+          <img
+            src={headshot}
+            alt={name}
+            style={{
+              width: headshotSize,
+              height: headshotSize,
+            }}
+          />
+          <div className="conversation">
+            <div className="greeting speech-bubble">
               <h1>Hi, I'm {name}</h1>
             </div>
-            <div className='action speech-bubble'>
-              <LinkedInButton/>
-              <TwitterButton/>
+            <div className="action speech-bubble">
+              <LinkedInButton />
+              <TwitterButton />
             </div>
           </div>
         </div>
