@@ -1,13 +1,13 @@
-import React from 'react';
-import LazyLoad from 'react-lazyload';
+import React from "react";
+import LazyLoad from "react-lazyload";
 
 export interface LogoProps {
   /** Name of the organisation or entity */
-  name: string,
+  name: string;
   /** Logo URL */
-  src: string,
+  src: string;
   /** URL to their website or a relevant page */
-  url: string,
+  url: string;
 }
 
 /**
@@ -18,9 +18,9 @@ export default class Logo extends React.Component<LogoProps> {
     const { name, src, url } = this.props;
 
     return (
-      <a className='logo' href={url} key={name}>
+      <a className="logo" href={url} key={name}>
         <LazyLoad height={150} offset={100}>
-          <img src={src} alt={name}/>
+          <img src={src} alt={name} />
         </LazyLoad>
       </a>
     );
