@@ -1,10 +1,10 @@
-import React from 'react';
-import { mount } from 'enzyme';
+import React from "react";
+import { render } from "@testing-library/react";
 
-import App from '../App';
+import App from "../App";
 
-describe('App', () => {
-  it('renders without crashing', () => {
-    mount(<App/>);
+describe("App", () => {
+  it("renders without crashing", () => {
+    render(<App />);
   });
 });
