@@ -1,21 +1,21 @@
-import React from 'react';
-import { mount } from 'enzyme';
-import renderer from 'react-test-renderer';
+import React from "react";
+import { render } from "@testing-library/react";
+import renderer from "react-test-renderer";
 
-import Showcase, { ShowcaseProps } from '../components/Showcase';
+import Showcase, { ShowcaseProps } from "../components/Showcase";
 
 const props: ShowcaseProps = {
-  title: 'title',
+  title: "title",
   children: <div></div>,
 };
 
-describe('Showcase', () => {
-  it('renders without crashing', () => {
-    mount(<Showcase {...props}/>);
+describe("Showcase", () => {
+  it("renders without crashing", () => {
+    render(<Showcase {...props} />);
   });
 
-  it('should match the snapshot', () => {
-    const tree = renderer.create(<Showcase {...props}/>).toJSON();
+  it("should match the snapshot", () => {
+    const tree = renderer.create(<Showcase {...props} />).toJSON();
     expect(tree).toMatchSnapshot();
   });
 });
